Force use short

+++ src/Ex5.java
@@ -3,8 +3,8 @@ import java.util.Calendar;
 public class Ex5 {
     public static void main(String[] args) {
         byte age =28;
-        int currYear = Calendar.getInstance().get(Calendar.YEAR);
-        int yearOfBirth = currYear - age;
+        short currYear = (short) (Calendar.getInstance().get(Calendar.YEAR));
+        short yearOfBirth = (short) (currYear - age);
         System.out.println("Year of birth = " + yearOfBirth);
     }
 }
